refactor(server): Move ast subpackage to zally-core #560
