+++ server/src/main/java/de/zalando/zally/rule/JsonRulesValidator.kt
@@ -3,7 +3,7 @@ package de.zalando.zally.rule
 import com.fasterxml.jackson.core.JsonPointer
 import com.fasterxml.jackson.databind.JsonNode
 import de.zalando.zally.core.ObjectTreeReader
-import de.zalando.zally.util.ast.ReverseAst
+import de.zalando.zally.core.ast.ReverseAst
 import io.swagger.util.Json
 
 class JsonRulesValidator(rules: RulesManager) : RulesValidator<JsonNode>(rules) {
