+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/ast/ReverseAst.kt
@@ -1,4 +1,4 @@
-package de.zalando.zally.util.ast
+package de.zalando.zally.core.ast
 
 import com.fasterxml.jackson.core.JsonPointer
