+++ server/zally-core/src/test/kotlin/de/zalando/zally/core/ast/ReverseAstBuilderTest.kt
@@ -1,6 +1,6 @@
-package de.zalando.zally.util.ast
+package de.zalando.zally.core.ast
 
-import de.zalando.zally.util.ast.ReverseAstBuilder.Companion.traversalMethods
+import de.zalando.zally.core.ast.ReverseAstBuilder.Companion.traversalMethods
 import io.swagger.models.Swagger
 import io.swagger.v3.oas.models.OpenAPI
 import io.swagger.v3.oas.models.PathItem
